website project upload
Introducing the Student Study Planner Website!
This website is designed to be your go-to tool for staying on top of your study schedule. With a simple and intuitive interface, you'll be able to easily plan and organize your tasks with our study planner table.

- A study planner table that lets you see your tasks by day and time
- A section for your user information, including a profile picture, name, and interests
- A fast and reliable user experience, thanks to our static website built with HTML and CSS

My goal is to provide you with a simple yet effective tool to manage your study schedule. We want to help you stay organized, focused, and achieve your academic goals!
I wish to make upgrades and increase the function when I have proper resources